+++ test/services/Reports.js
@@ -2,14 +2,59 @@ var expect = require('chai').expect;
 var helpers = require('../helpers');
 var api = require('../main');
 
-describe('Reports - Transaction', function() {
+describe('Reports - Transactions', function () {
     var report;
 
-    before(function(done) {
+    before(function (done) {
         report = {
-            ReportType: "TRANSACTION"
+            ReportType: "TRANSACTIONS",
+            Tag: 'Created with Mangopay NodeJs SDK',
+            DownloadFormat: 'CSV',
+            CallbackURL: 'https://mangopay.com/docs/please-ignore',
+            Sort: 'CreationDate:ASC',
+            Preview: false,
+            Filters:
+                {
+                    BeforeDate: 1658838931,
+                    AfterDate: 1658838931,
+                    Type: ['PAYIN'],
+                    ResultCode: ['000000'],
+                    Status: ['SUCCEEDED'],
+                    Nature: ['REGULAR'],
+                    WalletId: null,
+                    AuthorId: null,
+                    MinDebitedFundsAmount: 10,
+                    MinDebitedFundsCurrency: 'EUR',
+                    MaxDebitedFundsAmount: 12000,
+                    MaxDebitedFundsCurrency: 'EUR',
+                    MinFeesAmount: 10,
+                    MinFeesCurrency: 'EUR',
+                    MaxFeesAmount: 150000,
+                    MaxFeesCurrency: 'EUR',
+                },
+            Columns: [
+                'Id',
+                'Tag',
+                'CreationDate',
+                'ExecutionDate',
+                'AuthorId',
+                'CreditedUserId',
+                'DebitedFundsAmount',
+                'DebitedFundsCurrency',
+                'CreditedFundsAmount',
+                'CreditedFundsCurrency',
+                'FeesAmount',
+                'FeesCurrency',
+                'Status',
+                'ResultCode',
+                'ResultMessage',
+                'Type',
+                'Nature',
+                'CreditedWalletId',
+                'DebitedWalletId',
+            ]
         };
-        api.Reports.create(report).then(function () {
+        api.Reports.create(report).then(function (data) {
             done();
         });
     });
@@ -21,8 +66,8 @@ describe('Reports - Transaction', function() {
     describe('Getting created report', function () {
         var getReport;
 
-        before(function(done){
-            api.Reports.get(report.Id).then(function(data){
+        before(function (done) {
+            api.Reports.get(report.Id).then(function (data) {
                 getReport = data;
                 done();
             });
@@ -35,12 +80,12 @@ describe('Reports - Transaction', function() {
     });
 });
 
-describe('Reports - Wallet', function() {
+describe('Reports - Wallets', function () {
     var report;
 
-    before(function(done) {
+    before(function (done) {
         report = {
-            ReportType: "WALLET"
+            ReportType: "WALLETS"
         };
         api.Reports.create(report).then(function () {
             done();
@@ -54,8 +99,8 @@ describe('Reports - Wallet', function() {
     describe('Getting created report', function () {
         var getReport;
 
-        before(function(done){
-            api.Reports.get(report.Id).then(function(data){
+        before(function (done) {
+            api.Reports.get(report.Id).then(function (data) {
                 getReport = data;
                 done();
             });