+++ typings/mangopay2-nodejs-sdk-tests.ts
@@ -41,14 +41,6 @@ const legalUser = new api.models.UserLegal({
         PostalCode: "20007",
         Country: "US"
     }),
-    LegalRepresentativeAddress: new api.models.Address({
-        AddressLine1: "4101 Reservoir Rd NW",
-        AddressLine2: "",
-        City: "Washington",
-        Region: "District of Columbia",
-        PostalCode: "20007",
-        Country: "US"
-    }),
     LegalRepresentativeBirthday: 1300186358,
     LegalRepresentativeNationality: "FR",
     LegalRepresentativeCountryOfResidence: "FR",
@@ -179,15 +171,7 @@ const userLegalPayer = new api.models.UserLegalPayer({
     LegalPersonType: "BUSINESS",
     LegalRepresentativeFirstName: "Mango",
     LegalRepresentativeLastName: "Pay",
-    LegalRepresentativeEmail: "mango@mangopay.com",
-    LegalRepresentativeAddress: new api.models.Address({
-        AddressLine1: "4101 Reservoir Rd NW",
-        AddressLine2: "",
-        City: "Washington",
-        Region: "District of Columbia",
-        PostalCode: "20007",
-        Country: "US"
-    })
+    LegalRepresentativeEmail: "mango@mangopay.com"
 });
 
 api.Users.create(userLegalPayer,
@@ -205,14 +189,6 @@ const userLegalOwner = new api.models.UserLegalOwner({
     LegalRepresentativeFirstName: "Mango",
     LegalRepresentativeLastName: "Pay",
     HeadquartersAddress: new api.models.Address({
-        AddressLine1: "4101 Reservoir Rd NW",
-        AddressLine2: "",
-        City: "Washington",
-        Region: "District of Columbia",
-        PostalCode: "20007",
-        Country: "US"
-    }),
-    LegalRepresentativeAddress: new api.models.Address({
         AddressLine1: "4101 Reservoir Rd NW",
         AddressLine2: "",
         City: "Washington",
@@ -626,7 +602,7 @@ api.PayIns.create({
             PostalCode: "68400"
         }
     },
-    MerchantOrderId: "1234"
+    Reference: "1234"
 }).then(data => {
     const d = data; // $ExpectType KlarnaWebPayInData
 });
@@ -675,6 +651,50 @@ api.PayIns.create({
 });
 
 api.PayIns.create({
+    PaymentType: "IDEAL",
+    ExecutionType: "WEB",
+    AuthorId: "user-id",
+    CreditedWalletId: "wallet-id",
+    Fees: {Amount: 100, Currency: "GBP"},
+    DebitedFunds: {Amount: 2000, Currency: "GBP"},
+    ReturnURL: "http://test.com",
+    StatementDescriptor: "Ideal",
+    Bic: "RBRBNL21",
+    Tag: "test"
+}).then(data => {
+    const d = data; // $ExpectType IdealWebPayInData
+});
+
+api.PayIns.create({
+    PaymentType: "GIROPAY",
+    ExecutionType: "WEB",
+    AuthorId: "user-id",
+    CreditedWalletId: "wallet-id",
+    Fees: {Amount: 100, Currency: "GBP"},
+    DebitedFunds: {Amount: 2000, Currency: "GBP"},
+    ReturnURL: "http://test.com",
+    StatementDescriptor: "Giropay",
+    Tag: "test"
+}).then(data => {
+    const d = data; // $ExpectType GiropayWebPayInData
+});
+
+api.PayIns.create({
+    PaymentType: "CARD",
+    ExecutionType: "WEB",
+    AuthorId: "user-id",
+    CreditedWalletId: "wallet-id",
+    Fees: {Amount: 100, Currency: "GBP"},
+    DebitedFunds: {Amount: 2000, Currency: "GBP"},
+    ReturnURL: "https://secure-return.co",
+    Culture: "FR",
+    Bic: "RBRBNL21",
+    CardType: "IDEAL"
+}).then(data => {
+    const d = data; // $ExpectType CardWebPayInData
+});
+
+api.PayIns.create({
     PaymentType: "CARD",
     ExecutionType: "WEB",
     AuthorId: "user-id",
@@ -719,6 +739,7 @@ api.PayIns.createPayPal({
     ReturnURL: "http://test.com",
     Tag: "test tag",
     StatementDescriptor: "test",
+    Reference: "Reference",
     Culture: "FR"
 }).then(data => {
     const d = data; // $ExpectType PayPalWebPayInData
@@ -808,6 +829,14 @@ api.PayIns.get("payin-id").then(data => {
 });
 
 api.PayIns.createRefund("payin-id", {AuthorId: "user-id"}).then(data => {
+    const d = data; // $ExpectType RefundData
+});
+
+api.PayIns.createRefund("payin-id", {
+    AuthorId: "user-id",
+    DebitedFunds: {Amount: 100, Currency: "EUR"},
+    Fees: {Amount: 15, Currency: "EUR"}
+}).then(data => {
     const d = data; // $ExpectType RefundData
 });
 
@@ -1014,7 +1043,9 @@ api.Transfers.getRefunds("transfer-id").then(data => {
 
 api.BankingAliases.create({
     Country: "GB",
-    OwnerName: "owner-id"
+    OwnerName: "owner-id",
+    Type: 'IBAN',
+    WalletId: '1234'
 }).then(data => {
     const d = data; // $ExpectType IBANBankingAliasData
 });
@@ -1024,7 +1055,10 @@ api.BankingAliases.get("alias-id").then(data => {
 api.BankingAliases.getAll().then(data => {
     const d = data; // $ExpectType IBANBankingAliasData[]
 });
-api.BankingAliases.update({OwnerName: "some-name"}).then(data => {
+api.BankingAliases.update({
+    Id: '1234',
+    Active: false
+}).then(data => {
     const d = data; // $ExpectType IBANBankingAliasData
 });
 api.BankingAliases.deactivate("alias-id").then(data => {
@@ -1222,7 +1256,7 @@ api.Hooks.getAll().then(data => {
 
 /* Reports */
 
-api.Reports.create({Columns: ["Alias", "AuthorId"], ReportType: "WALLET"}).then(data => {
+api.Reports.create({Columns: ["Alias", "AuthorId"], ReportType: "WALLETS"}).then(data => {
     const d = data; // $ExpectType ReportData
 });
 
@@ -1292,6 +1326,34 @@ api.PayIns.createCardPreAuthorizedDepositPayIn({
 }, data => {
     const d = data; // $ExpectType CardPreAuthorizedDepositPayInData
 });
+
+api.InstantConversions.getConversionRate("EUR", "GBP")
+    .then(data => {
+        const d = data; // $ExpectType ConversionRateData
+    });
+
+api.InstantConversions.createInstantConversion(
+    {
+        AuthorId: "author-id",
+        CreditedWalletId: "credited-wallet-id",
+        DebitedWalletId: "debited-wallet-id",
+        CreditedFunds: {
+            Amount: 0,
+            Currency: "GBP"
+        },
+        DebitedFunds: {
+            Amount: 79,
+            Currency: "EUR"
+        }
+    }
+).then(data => {
+    const d = data; // $ExpectType InstantConversionData
+});
+
+api.InstantConversions.getInstantConversion("conversion-id")
+    .then(data => {
+        const d = data; // $ExpectType InstantConversionData
+    });
 
 /* Namespace Access */